+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".core.features.landing.presentation.LandingActivity"
+            android:name=".features.landing.presentation.LandingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -29,6 +29,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <meta-data
+            android:name="preloaded_fonts"
+            android:resource="@array/preloaded_fonts" />
     </application>
 
 </manifest>
