+++ app/src/main/java/com/online/template/core/data/local/database/DataBase.kt
@@ -3,8 +3,8 @@ package com.online.template.core.data.local.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.online.template.core.features.landing.TestDao
-import com.online.template.core.features.landing.TestDataClass
+import com.online.template.features.TestDao
+import com.online.template.features.TestDataClass
 
 @Database(entities = [TestDataClass::class], version = 1000, exportSchema = false)
 @TypeConverters()
