--- app/src/main/java/com/online/template/core/features/landing/data/remote/data_sources/RemoteDataSourceContract.kt
@@ -1,7 +1,0 @@
-package com.online.template.core.features.landing.data.remote.data_sources
-
-import retrofit2.Response
-
-interface RemoteDataSourceContract {
-    suspend fun getTest(): Response<Unit>
-}
