--- app/src/main/java/com/online/template/core/features/landing/domain/abstraction/LandingRepositoryContract.kt
@@ -1,5 +1,0 @@
-package com.online.template.core.features.landing.domain.abstraction
-
-interface LandingRepositoryContract {
-    suspend fun getTest(): Unit
-}
