--- app/src/main/java/com/online/template/core/features/landing/presentation/LandingActivity.kt
@@ -1,21 +1,0 @@
-package com.online.template.core.features.landing.presentation
-
-import android.os.Bundle
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import com.online.template.R
-import com.online.template.databinding.ActivityLandingBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class LandingActivity : AppCompatActivity() {
-    private var binding: ActivityLandingBinding? = null
-    private val viewModel: LandingViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_landing)
-        binding?.textViewLanding?.setOnClickListener { viewModel.start() }
-    }
-}
