+++ app/src/main/java/com/online/template/features/landing/data/remote/data_sources/RemoteDataSourceContract.kt
@@ -1,0 +1,7 @@
+package com.online.template.features.landing.data.remote.data_sources
+
+import retrofit2.Response
+
+interface RemoteDataSourceContract {
+    suspend fun getTest(): Response<Unit>
+}
