+++ app/src/main/java/com/online/template/features/landing/data/remote/services/LandingService.kt
@@ -1,0 +1,9 @@
+package com.online.template.features.landing.data.remote.services
+
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface LandingService {
+    @GET("/c/3029-d29f-4014-9fb4")
+    suspend fun testRetrofit(): Response<Unit>
+}
