+++ app/src/main/java/com/online/template/features/landing/domain/abstraction/LandingRepositoryContract.kt
@@ -1,0 +1,5 @@
+package com.online.template.features.landing.domain.abstraction
+
+interface LandingRepositoryContract {
+    suspend fun getTest(): Unit
+}
