+++ app/src/main/java/com/online/template/features/landing/presentation/LandingActivity.kt
@@ -1,0 +1,20 @@
+package com.online.template.features.landing.presentation
+
+import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.online.template.R
+import com.online.template.databinding.ActivityLandingBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LandingActivity : AppCompatActivity() {
+    private var binding: ActivityLandingBinding? = null
+    private val viewModel: LandingViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_landing)
+    }
+}
