+++ app/src/main/java/com/online/template/features/landing/presentation/main/LandingMainViewModel.kt
@@ -1,0 +1,6 @@
+package com.online.template.features.landing.presentation.main
+
+import com.online.template.core.presentation.activity.CoreViewModel
+import dagger.assisted.AssistedInject
+
+class LandingMainViewModel @AssistedInject constructor() : CoreViewModel()
