+++ app/src/main/java/com/online/template/features/landing/presentation/main/LandingMainViewModelFactory.kt
@@ -1,0 +1,8 @@
+package com.online.template.features.landing.presentation.main
+
+import dagger.assisted.AssistedFactory
+
+@AssistedFactory
+interface LandingMainViewModelFactory {
+    fun create(): LandingMainViewModel
+}
