+++ app/src/main/java/com/online/template/features/landing/presentation/registration/RegistrationViewModel.kt
@@ -1,0 +1,6 @@
+package com.online.template.features.landing.presentation.registration
+
+import com.online.template.core.presentation.activity.CoreViewModel
+import dagger.assisted.AssistedInject
+
+class RegistrationViewModel @AssistedInject constructor(): CoreViewModel()
