+++ app/src/main/java/com/online/template/features/landing/presentation/registration/RegistrationViewModelFactory.kt
@@ -1,0 +1,9 @@
+package com.online.template.features.landing.presentation.registration
+
+import com.online.template.features.landing.presentation.registration.RegistrationViewModel
+import dagger.assisted.AssistedFactory
+
+@AssistedFactory
+interface RegistrationViewModelFactory {
+    fun create(): RegistrationViewModel
+}
