+++ app/src/main/res/layout/activity_landing.xml
@@ -9,14 +9,12 @@
         android:layout_height="match_parent"
         tools:context=".features.landing.presentation.LandingActivity">
 
-        <TextView
-            android:id="@+id/textViewLanding"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Landing"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fragmentContainerViewLanding"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:navGraph="@navigation/landing_graph"
+            app:defaultNavHost="true"
+            android:name="androidx.navigation.fragment.NavHostFragment"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
