+++ app/src/main/res/navigation/landing_graph.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/landing_graph"
+    app:startDestination="@id/registrationFragment">
+
+    <fragment
+        android:id="@+id/registrationFragment"
+        android:name="com.online.template.features.landing.presentation.registration.RegistrationFragment"
+        android:label="RegistrationFragment"
+        tools:layout="@layout/fragment_registration" />
+    <fragment
+        android:id="@+id/landingMainFragment"
+        android:name="com.online.template.features.landing.presentation.main.LandingMainFragment"
+        android:label="LandingMainFragment"
+        tools:layout="@layout/fragment_landing_main" />
+</navigation>
