+++ app/src/main/res/values-en/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="donate_blood">Donate Blood</string>
+</resources>
