+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,8 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="primary_color">#800000</color>
+    <color name="background_color">#FFFFFFFF</color>
+    <color name="text_color">#000000</color>
+    <color name="secondary_text_color">#827D7D</color>
 </resources>
