+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Template</string>
+    <string name="donate_blood">Doniraj Krv</string>
 </resources>
